requirements.txt
numpy
pandas
requests
scikit-learn
matplotlib
seaborn
joblib
statsmodels
scipy
pyyaml
streamlit
plotly
